Set the flush threshold on a per-segment basis
Set the LLC realtime segment flush threshold on a per-segment basis
according to the maximum number of partitions consumed by replicas. For
example, if some replicas for a new segment are consuming three
partitions and some are consuming four, the new segment will have a
flush threshold equal to the per-table row flush threshold divided by
four.

This ensures that even if the number of Kafka partitions increases,
there is still an upper bound on the number of rows held in memory
before being flushed to disk.